Fix typo
Without this patch, if a test has been registered, it may NOT be retrived
from the current pool. Now, we ensure the test is used.

Without this patch, a project test suite will stop in the middle,
because some tests would never be finished (because tests are
duplicated) so one parent test is finished whereas its copy is not.